list component rename
