+++ templates/Web/Pages/Angular.List/src/app/app-shell/AngularList/list-form/list-form.component.html
