+++ templates/Web/Pages/Angular.List/src/app/app-shell/AngularList/list-item/list-item.ts
